+++ src/test/kotlin/no/nav/sbl/sosialhjelpmodiaapi/service/soknadsstatus/SoknadsStatusServiceTest.kt
@@ -1,0 +1,44 @@
+package no.nav.sbl.sosialhjelpmodiaapi.service.soknadsstatus
+
+import io.mockk.clearAllMocks
+import io.mockk.every
+import io.mockk.mockk
+import no.nav.sbl.sosialhjelpmodiaapi.domain.DigisosSak
+import no.nav.sbl.sosialhjelpmodiaapi.domain.InternalDigisosSoker
+import no.nav.sbl.sosialhjelpmodiaapi.domain.SoknadsStatus
+import no.nav.sbl.sosialhjelpmodiaapi.domain.SoknadsStatusResponse
+import no.nav.sbl.sosialhjelpmodiaapi.event.EventService
+import no.nav.sbl.sosialhjelpmodiaapi.client.fiks.FiksClient
+import org.assertj.core.api.Assertions.assertThat
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
+
+internal class SoknadsStatusServiceTest {
+
+    private val fiksClient: FiksClient = mockk()
+    private val eventService: EventService = mockk()
+
+    private val service = SoknadsStatusService(fiksClient, eventService)
+
+    private val mockDigisosSak: DigisosSak = mockk()
+    private val mockInternalDigisosSoker: InternalDigisosSoker = mockk()
+
+    private val token = "token"
+
+    @BeforeEach
+    fun init() {
+        clearAllMocks()
+        every { fiksClient.hentDigisosSak(any(), any()) } returns mockDigisosSak
+    }
+
+    @Test
+    fun `Skal returnere mest nylige SoknadsStatus`() {
+        every { eventService.createModel(any(), any()) } returns mockInternalDigisosSoker
+        every { mockInternalDigisosSoker.status } returns SoknadsStatus.UNDER_BEHANDLING
+
+        val response: SoknadsStatusResponse = service.hentSoknadsStatus("123", token)
+
+        assertThat(response).isNotNull
+        assertThat(response.status).isEqualTo(SoknadsStatus.UNDER_BEHANDLING)
+    }
+}
